python_version/requirements.txt
# Dependências do Python
paho-mqtt
requests